moved student application to relation
